requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
aiosqlite>=0.19.0
httpx>=0.25.1
pytest>=7.4.3
websockets>=12.0  # Required for WebSocket support
typing_extensions>=4.8.0  # Required for Python 3.7+ type hints
